Move all plugins stuff to lib/rollbar/plugins
The middlewares are still in lib/rollbar/middlewares, but everything is
now loaded from the plugins directory
