--- lib/rollbar/core_ext/basic_socket.rb
@@ -1,7 +1,0 @@
-require 'socket'
-
-class BasicSocket
-  def as_json
-    to_s
-  end
-end
