--- lib/rollbar/js/frameworks.rb
@@ -1,6 +1,0 @@
-module Rollbar
-  module Js
-    module Frameworks
-    end
-  end
-end
