--- lib/rollbar/js/version.rb
@@ -1,5 +1,0 @@
-module Rollbar
-  module Js
-    VERSION = "0.1.0"
-  end
-end
