+++ lib/rollbar/plugins/active_job.rb
