+++ lib/rollbar/plugins/rails/controller_methods.rb
