+++ spec/rollbar/plugins/delayed_job/job_data.rb
