+++ spec/rollbar/plugins/rails_js_spec.rb
@@ -11,7 +11,7 @@ describe ApplicationController, :type => 'request' do
   it 'renders the snippet and config in the response', :type => 'request' do
     get '/test_rollbar_js'
 
-    snippet_from_submodule = File.read(File.expand_path('../../../../../rollbar.js/dist/rollbar.snippet.js', __FILE__))
+    snippet_from_submodule = File.read(File.expand_path('../../../../rollbar.js/dist/rollbar.snippet.js', __FILE__))
 
     expect(response.body).to include("var _rollbarConfig = #{Rollbar::configuration.js_options.to_json};")
     expect(response.body).to include(snippet_from_submodule)
